# user-server/docs/swagger.yaml
basePath: /
definitions:
...
  hivemtk-user_internal_pkg_utils_response.Response:
    properties:
      code: {}
      data: {}
      message:
        type: string
    type: object
...
  hivemtk-user_internal_service.MergePreviewRule:
    properties:
      enabled:
        type: boolean
      field:
        type: string
      fields:
        items:
          type: string
        type: array
      name:
        type: string
      op:
        type: string
      threshold:
        type: integer
    type: object
  hivemtk-user_internal_service.MergeRule:
    properties:
      built_in:
        type: boolean
      description:
        type: string
      enabled:
        type: boolean
      example:
        type: string
      field:
        type: string
      id:
        type: integer
      name:
        type: string
      op:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
      value:
        type: string
    type: object
  hivemtk-user_internal_service.MergeRuleSet:
    properties:
      built_in:
        items:
          $ref: '#/definitions/hivemtk-user_internal_service.MergeRule'
        type: array
      custom:
        items:
          $ref: '#/definitions/hivemtk-user_internal_service.MergeRule'
        type: array
      strategy:
        $ref: '#/definitions/hivemtk-user_internal_service.MergeStrategy'
      updated_at:
        type: string
    type: object
  hivemtk-user_internal_service.MergeStrategy:
    properties:
      conflict_behavior:
        type: string
      post_merge_actions:
        items:
          type: string
        type: array
      primary_rule:
        type: string
      window_end:
        type: string
      window_start:
        type: string
    type: object
...
  internal_controller.mergeRequest:
    properties:
      primary_id:
        type: string
      secondary_id:
        type: string
    required:
    - primary_id
    - secondary_id
    type: object
  internal_service.MergeRule:
    properties:
      built_in:
        type: boolean
      description:
        type: string
      enabled:
        type: boolean
      example:
        type: string
      field:
        type: string
      id:
        type: integer
      name:
        type: string
      op:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
      value:
        type: string
    type: object
  internal_service.MergeRuleSet:
    properties:
      built_in:
        items:
          $ref: '#/definitions/internal_service.MergeRule'
        type: array
      custom:
        items:
          $ref: '#/definitions/internal_service.MergeRule'
        type: array
      strategy:
        $ref: '#/definitions/internal_service.MergeStrategy'
      updated_at:
        type: string
    type: object
  internal_service.MergeStrategy:
    properties:
      conflict_behavior:
        type: string
      post_merge_actions:
        items:
          type: string
        type: array
      primary_rule:
        type: string
      window_end:
        type: string
      window_start:
        type: string
    type: object
info:
  contact: {}
  description: |-
    HiveMtk 用户端（私域独立部署）RESTful API。认证采用 JWT：请求头 Authorization: Bearer <token>。
    仅当 ENABLE_SWAGGER=true 时，通过 http://localhost:8204/swagger/index.html 本地访问。
  title: HiveMtk 用户端 API
  version: 3.41.0
paths:
...
  /api/oneid:
    get:
      description: 分页查询客户 OneID 及其身份
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页
        in: query
        name: page_size
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: OneID 列表
      tags:
      - OneID
  /api/oneid/merge:
    post:
      consumes:
      - application/json
      description: 把两个 OneID 合并为一个，保留主 ID 全部行为轨迹
      parameters:
      - description: 合并请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.mergeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 合并成功
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: 合并 OneID 客户身份
      tags:
      - OneID
  /api/oneid/merge-rules:
    get:
      description: 返回预置规则 + 自定义规则 + 合并策略
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_service.MergeRuleSet'
      security:
      - BearerAuth: []
      summary: 获取 OneID 合并规则集
      tags:
      - OneID
    post:
      consumes:
      - application/json
      description: 全量替换当前规则集；预置规则只允许切换 enabled / 调整 priority
      parameters:
      - description: 规则集
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hivemtk-user_internal_service.MergeRuleSet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hivemtk-user_internal_service.MergeRuleSet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: 保存 OneID 合并规则集
      tags:
      - OneID
  /api/oneid/merge-rules/preview:
    post:
      consumes:
      - application/json
      description: 按提交的规则集预览「会合并哪些身份」，返回候选合并对数量与样例
      parameters:
      - description: 规则数组（前端 MergeRuleConfig.vue 扁平结构）
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/hivemtk-user_internal_service.MergePreviewRule'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 'data: {candidateCount, samples:[{from,to,score}]}'
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: 合并规则命中预览
      tags:
      - OneID
  /api/oneid/stats:
    get:
      description: 返回 OneID 总数、冲突数、合并率等
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/hivemtk-user_internal_pkg_utils_response.Response'
      security:
      - BearerAuth: []
      summary: OneID 统计
      tags:
      - OneID
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"